requirements.txt
flask
kafka-python
pymongo
docker